change fetch home feed
