--- app/loading.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import { LeapFrog } from '@uiball/loaders'
-
-function Loading() {
-  return (
-    <LeapFrog
-    size={40}
-    speed={2.5}
-    color="black"
-    />
-  )
-}
-
-export default Loading
