+++ app/page.jsx
@@ -1,6 +1,22 @@
+'use client'
+
 import Feed from "@components/Feed"
+import { useState, useEffect } from "react"
 
 const Home = () => {
+  const [posts, setPosts] = useState([])
+
+  useEffect(() => {
+    const fetchPosts = async () => {
+      const response = await fetch('/api/tool')
+      const data = await response.json()
+  
+      setPosts(data)
+    }
+    
+    fetchPosts()
+  }, [])
+
   return (
     <section className="w-full flex-center flex-col">
         <h1 className="head_text text-center">
@@ -11,7 +27,10 @@ const Home = () => {
         <p className="desc text-center">
             ToolBox is an open-source database of the most useful tools on the internet
         </p>
-        <Feed />
+        <Feed
+          posts={posts}
+          
+        />
     </section>
   )
 }
